Fixing io for cube format if data is None.
There exist a terrible hack in this code: If the shape of the input
data is odd numbers, it is assumed that it is gpaw data with
non-periodic boundary conditions. Therefore some fixes are made to
shift the origin of the volumetric data.

This should not be done if data is None, which is why

if data is None:
  data = np.ones(1, 1, 1)

is now replaced by 
if data is None:
  data = np.ones(2, 2, 2)

So this "fix" is avoided!