Created a 'run' method. This makes the code more readable.
Also added output_template and track_output keywords. If the 
environmental parameter VASP_COMMAND is used to execute VASP,
the stdout from VASP is now written to a file (default vasp.out) 
instead of to the stdout of ASE. If VASP_SCRIPT is used the VASP stdout, 
is still printed as stdout for the ASE run (can this be done also for 
VASP_SCRIPT?).