NWChem: issue warning if orbitals switched
If the gap between HOMO and LUMO is too small,
NWChem raises the energy of the LUMO. If the orbitals
are degenerate this lead to an artificial break in 
symmetry which can cause false results.

So if we observe a level-switch in the results,
we inform the user that this result might be false.